move makefile stuff around
